rename MqClientImpl to MqConnection
