+++ kas-mq-admin/src/com/kas/mq/admin/commands/ConnectCommand.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import java.util.TreeSet;
 import com.kas.infra.typedef.TokenDeque;
 import com.kas.infra.utils.Validators;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 
 /**
  * A CONNECT command
@@ -29,7 +29,7 @@ public class ConnectCommand extends ACliCommand
    * @param args The command arguments specified when command was entered
    * @param client The client that will perform the actual connection
    */
-  protected ConnectCommand(Scanner scanner, TokenDeque args, MqClientImpl client)
+  protected ConnectCommand(Scanner scanner, TokenDeque args, MqConnection client)
   {
     super(scanner, args, client);
   }
