+++ kas-mq-admin/src/com/kas/mq/admin/commands/DefQueueCommand.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 import com.kas.infra.typedef.TokenDeque;
 import com.kas.infra.utils.Validators;
 import com.kas.mq.impl.internal.IMqConstants;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 
 /**
  * A DEFINE QUEUE command
@@ -21,7 +21,7 @@ public class DefQueueCommand extends ACliCommand
    * @param args The command arguments specified when command was entered
    * @param client The client that will perform the actual connection
    */
-  protected DefQueueCommand(Scanner scanner, TokenDeque args, MqClientImpl client)
+  protected DefQueueCommand(Scanner scanner, TokenDeque args, MqConnection client)
   {
     super(scanner, args, client);
   }
