+++ kas-mq-admin/src/com/kas/mq/admin/commands/DelQueueCommand.java
@@ -3,7 +3,7 @@ package com.kas.mq.admin.commands;
 import java.util.Scanner;
 import com.kas.infra.typedef.TokenDeque;
 import com.kas.infra.utils.Validators;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 
 /**
  * A DELETE command
@@ -20,7 +20,7 @@ public class DelQueueCommand extends ACliCommand
    * @param args The command arguments specified when command was entered
    * @param client The client that will perform the actual connection
    */
-  protected DelQueueCommand(Scanner scanner, TokenDeque args, MqClientImpl client)
+  protected DelQueueCommand(Scanner scanner, TokenDeque args, MqConnection client)
   {
     super(scanner, args, client);
   }
