+++ kas-mq-admin/src/com/kas/mq/admin/commands/DisconnectCommand.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 import java.util.Set;
 import java.util.TreeSet;
 import com.kas.infra.typedef.TokenDeque;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 
 /**
  * A DISCONNECT command
@@ -28,7 +28,7 @@ public class DisconnectCommand extends ACliCommand
    * @param args The command arguments specified when command was entered
    * @param client The client that will perform the actual disconnection
    */
-  protected DisconnectCommand(Scanner scanner, TokenDeque args, MqClientImpl client)
+  protected DisconnectCommand(Scanner scanner, TokenDeque args, MqConnection client)
   {
     super(scanner, args, client);
   }
