+++ kas-mq-admin/src/com/kas/mq/admin/commands/ShutdownCommand.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 import java.util.Set;
 import java.util.TreeSet;
 import com.kas.infra.typedef.TokenDeque;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 
 /**
  * A SHUTDOWN command
@@ -27,7 +27,7 @@ public class ShutdownCommand extends ACliCommand
    * @param args The command arguments specified when command was entered
    * @param client The client that will perform the actual connection
    */
-  protected ShutdownCommand(Scanner scanner, TokenDeque args, MqClientImpl client)
+  protected ShutdownCommand(Scanner scanner, TokenDeque args, MqConnection client)
   {
     super(scanner, args, client);
   }
