+++ kas-mq-server/src/com/kas/mq/server/processors/DeleteQueueProcessor.java
@@ -5,7 +5,7 @@ import com.kas.comm.impl.NetworkAddress;
 import com.kas.mq.impl.IMqMessage;
 import com.kas.mq.impl.internal.EMqCode;
 import com.kas.mq.impl.internal.IMqConstants;
-import com.kas.mq.impl.internal.MqClientImpl;
+import com.kas.mq.impl.internal.MqConnection;
 import com.kas.mq.impl.internal.MqLocalQueue;
 import com.kas.mq.server.IController;
 import com.kas.mq.server.IRepository;
@@ -119,7 +119,7 @@ public class DeleteQueueProcessor extends AProcessor
         
         mLogger.debug("DeleteQueueProcessor::postprocess() - Notifying KAS/MQ server \"" + remoteQmgrName + "\" (" + address.toString() + ") on repository update");
         
-        MqClientImpl client = new MqClientImpl();
+        MqConnection client = new MqConnection();
         client.connect(address.getHost(), address.getPort());
         if (client.isConnected())
         {
